requirements.txt
google-cloud-secret-manager==2.0.0
Jinja2==2.10.3
psycopg2-binary==2.8.4
python-json-logger==0.1.11
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
SQLAlchemy==1.3.12
pandas_gbq
google-cloud-pubsub